requirements.txt
numpy==1.15.4
pandas==1.1.5
matplotlib==3.2.2
scikit-image==0.16.2
scikit-learn==0.23.2
jupyter==1.0.0
tensorflow-gpu==1.13.1
keras-gpu==2.3.1
python-mnist==0.7